Halve runtime on large lines
This dramatically improves performance on lines with a large number of points, primarily by reducing the number of times we have to iterate over the list.

- Removes calls to `length/1`, since it's linear in the size of the list
- Cuts down on the number of times we access the last element of a list
- Cuts down on the number of new lists we construct
- Cuts down on the number of times we iterate the full list to find the maximum distance

This also adds a very simple benchmark to the test suite, which can be run with:
$ mix test --only bench

On my machine (an M1 Mac running Elixir 13.4 + OTP 24, so no JIT support), the included benchmark runs in roughly 0.19 seconds per iteration in the shipping version of the code. With my changes applied, that drops down to roughly 0.08 seconds.